// .tmp/interfaces/options.interface.ts
export interface GridOptions {
    fields?: any;
    data?: any;
    autoload?: any;
    controller?: any;
    width?: any;
    height?: any;
    heading?: any;
    filtering?: any;
    inserting?: any;
    editing?: any;
    selecting?: any;
    sorting?: any;
    paging?: any;
    pageLoading?: any;
    rowClass?: any;
    rowClick?: any;
    rowDoubleClick?: any;
    noDataContent?: any;
    confirmDeleting?: any;
    deleteConfirm?: any;
    pagerContainer?: any;
    pageIndex?: any;
    pageSize?: any;
    pageButtonCount?: any;
    pagerFormat?: any;
    pageNextText?: any;
    pagePrevText?: any;
    pageFirstText?: any;
    pageLastText?: any;
    pageNavigatorNextText?: any;
    pageNavigatorPrevText?: any;
    invalidMessage?: any;
    invalidNotify?: any;
    loadIndication?: any;
    loadIndicationDelay?: any;
    loadMessage?: any;
    loadShading?: any;
    updateOnResize?: any;
    rowRenderer?: any;
    headerRowRenderer?: any;
    filterRowRenderer?: any;
    insertRowRenderer?: any;
    editRowRenderer?: any;
    pagerRenderer?: any;
    controlClass?: string;

    // Call-backs
    onDataLoading?: any;
    onDataLoaded?: any;
    onError?: any;
    onInit?: any;
    onItemDeleting?: any;
    onItemDeleted?: any;
    onItemEditing?: any;
    onItemInserting?: any;
    onItemInserted?: any;
    onItemUpdating?: any;
    onItemUpdated?: any;
    onItemInvalid?: any;
    onOptionChanging?: any;
    onOptionChanged?: any;
    onPageChanged?: any;
    onRefreshing?: any;
    onRefreshed?: any;
}